+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Backend and SPA Admin will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0](https://github.com/70000TONS-IT/booking-engine-admin/pull/785) - 2025-12-11
+
+### Added
+- More sophisticated survivor sync
+- Admin UI to manage potential matches
+- Enable running the command from the UI
 
 ## [1.5.5] - 2025-12-11